gh-665 improvements
